+++ src/app/errors/not-found-error.js
@@ -1,8 +1,17 @@
 'use strict';
 
-module.exports = class NotFoundError extends Error {
+const NOT_FOUND_HANDLER = (req, res) => {
+  res.status(404).send(`Page was not found`);
+};
+
+class NotFoundError extends Error {
   constructor(message) {
     super(message);
     this.code = 404;
   }
+}
+
+module.exports = {
+  NOT_FOUND_HANDLER,
+  NotFoundError
 };
